made fg working

+++ jobs_commands.h
@@ -14,6 +14,7 @@ ll convert(char *tokens)
 
 void fg(char **tokens)
 {
+	int og = 0;
 	if(tokens[2])
 	{
 		printf("Please provide the correct number of arguments\n");
@@ -21,10 +22,21 @@ void fg(char **tokens)
 	else if(!tokens[1])
 	{
 		printf("Please provide all arguments\n");
-		return; 
+		og = 1;
 	}
-	ll jno = convert(tokens[1]);
-	if(jno >= bgind){
+	printf("eji oji lo ji suno ji %lld\n",og);
+	ll jno ;
+	if(!og)
+	{
+		jno = convert(tokens[1]);
+	}
+	else
+	{
+		jno = bgind - 1;	
+	}
+	printf("%lld %s is the job number put into forefround\n",jno,characterarray[procaarray[jno]]);
+	if(jno >= bgind)
+	{
 		printf("Please provide correct job number\n");
 		return;
 	}
@@ -34,12 +46,89 @@ void fg(char **tokens)
 	}
 	bgind--;
 	pid_t pid = procaarray[jno];
+
+	/* Capture the terminal attributes in a struct termios file */
+	
+	struct termios term1,term2;
+	if(tcgetattr(0,&term1)!=0)
+	{
+		printf("Error transfering control\n");
+		return ;
+	}
+	/*
+	if(tcgetattr(1,&term2)!=0)
+	{
+		printf("Error transfering control\n");
+		return ;
+	}
+	*/
+
+	/* Give the controlling terminal to this process */
 	kill(pid,SIGCONT);
+	/*
+	if(tcsetpgrp(0,gidarray[jno])==-1)
+	{
+		perror("Error giving stdin\n");
+		return ;
+	}
+
+	if(tcsetpgrp(1,gidarray[jno])==-1)
+	{
+		perror("Error giving stdout\n");
+		return ;
+	}
+	*/
+	/* Give attributes to the process */
+	/*
+	if(tcsetattr(0,TCSNOW   ,&term1)!=0)
+	{
+		printf("Error transfering stdin\n");
+		return ;
+	}
+	if(tcsetattr(1,TCSNOW   ,&term2)!=0)
+	{
+		printf("Error transfering stdout\n");
+		return ;
+	}
+
+
+	printf("Gave stdout to process\n");
+	*/
+	printf("%ld ko bulaya\n",pid);
 	int status;
 	do
 	{
 		waitpid(pid,&status,WUNTRACED);
-	}while(!WIFEXITED(status) && !WIFSIGNALED(status) && !WIFSTOPPED(status));
+	}
+	while(!WIFEXITED(status) && !WIFSIGNALED(status) );
+	if(WIFEXITED(status))
+	{
+		printf("Lo\n");
+	}
+	else if(WIFSIGNALED(status))
+	{
+		printf("Fo\n");
+	}
+	else if(WIFSTOPPED(status))
+	{
+		printf("Go\n");
+	}
+
+	/* Return back controlling terminal from process to shell */
+	/*
+	tcsetpgrp(0,shell_gid);
+	tcsetpgrp(1,shell_gid);
+	if(tcsetattr(0,TCSNOW   ,&term1)!=0)
+	{
+		printf("Error transfering stdin\n");
+		return ;
+	}
+	if(tcsetattr(1,TCSNOW   ,&term2)!=0)
+	{
+		printf("Error transfering stdout\n");
+		return ;
+	}
+	*/
 	updatejobs();
 }
